bug fixed

+++ bot/handlers/mails.py
@@ -4,18 +4,35 @@ from bot.keyboards import bot_action, mail_action, gen_cancel, gen_ok
 from datetime import datetime
 
 
+async def safe_get_mail(uid: int, mail_id: int, cb_id: int | None = None) -> models.Mail | None:
+    try:
+        mail = mails_db.get(mail_id)
+        return mail
+    except data_exc.RecordIsMissing:
+        if cb_id:
+            await bot.answer_callback_query(
+                cb_id,
+                "❗️Помилка"
+            )
+        else:
+            await bot.send_message(
+                uid,
+                "❗️Помилка",
+                reply_markup=gen_ok("open_bot_list", "↩️Головне меню")
+            )
+        return None
+
+
 @dp.callback_query_handler(bot_action.filter(action="mails"), state="*")
-async def open_mail_list(cb: CallbackQuery, callback_data: dict):
+async def open_mail_list(cb: CallbackQuery, callback_data: dict, state: FSMContext):
+    await state.set_state(None)
     bot_dc = bots_db.get(int(callback_data["id"]))
     mails = mails_db.get_by(bot=int(callback_data["id"]))
     await cb.message.answer(
         "{text6}\nРозсилки:",
         reply_markup=kb.gen_mail_list(bot_dc, mails)
     )
-    try:
-        await cb.message.delete()
-    except MessageCantBeDeleted:
-        pass
+    await safe_del_msg(cb.from_user.id, cb.message.message_id)
 
 
 @dp.callback_query_handler(bot_action.filter(action="add_mail"))
@@ -32,10 +49,7 @@ async def add_mail(cb: CallbackQuery, callback_data: dict, state: FSMContext):
     )
     await state.set_state(states.InputStateGroup.mail)
     await state.set_data({"msg_id": msg.message_id, "bot_id": bot_dc.id})
-    try:
-        await cb.message.delete()
-    except MessageCantBeDeleted:
-        pass
+    await safe_del_msg(cb.from_user.id, cb.message.message_id)
 
 
 async def open_mail_menu(uid: int, mail_id: int, msg_id: int):
@@ -75,14 +89,14 @@ async def open_mail_menu(uid: int, mail_id: int, msg_id: int):
             mail.text,
             reply_markup=kb.gen_mail_menu(bot_dc, mail)
         )
-    try:
-        await bot.delete_message(uid, msg_id)
-    except MessageCantBeDeleted:
-        pass
+    await safe_del_msg(uid, msg_id)
 
 
-@dp.callback_query_handler(mail_action.filter(action="open_mail_menu"))
+@dp.callback_query_handler(mail_action.filter(action="open_mail_menu"), state="*")
 async def open_mail_menu_cb(cb: CallbackQuery, callback_data: dict):
+    mail = await safe_get_mail(cb.from_user.id, int(callback_data["id"]), cb.id)
+    if not mail:
+        return
     await open_mail_menu(cb.from_user.id, int(callback_data["id"]), cb.message.message_id)
 
 
@@ -100,10 +114,7 @@ async def add_mail(cb: CallbackQuery, callback_data: dict, state: FSMContext):
     )
     await state.set_state(states.InputStateGroup.mail)
     await state.set_data({"msg_id": msg.message_id, "bot_id": bot_dc.id})
-    try:
-        await cb.message.delete()
-    except MessageCantBeDeleted:
-        pass
+    await safe_del_msg(cb.from_user.id, cb.message.message_id)
 
 
 @dp.message_handler(content_types=ContentTypes.TEXT, state=states.InputStateGroup.mail)
@@ -167,6 +178,9 @@ async def mail_input_gif(msg: Message, state: FSMContext):
 
 @dp.callback_query_handler(mail_action.filter(action="add_buttons"))
 async def add_buttons(cb: CallbackQuery, callback_data: dict, state: FSMContext):
+    mail = await safe_get_mail(cb.from_user.id, int(callback_data["id"]), cb.id)
+    if not mail:
+        return
     msg = await cb.message.answer(
         "Щоб додати кнопки-посилання надішліть список у форматі\nпідпис1 - посилання1\nпідпис2 - посилання2\n...",
         reply_markup=gen_cancel(
@@ -176,10 +190,7 @@ async def add_buttons(cb: CallbackQuery, callback_data: dict, state: FSMContext)
             )
         )
     )
-    try:
-        await cb.message.delete()
-    except MessageCantBeDeleted:
-        pass
+    await safe_del_msg(cb.from_user.id, cb.message.message_id)
     await state.set_state(states.InputStateGroup.mail_buttons)
     await state.set_data({"msg_id": msg.message_id, "mail_id": callback_data["id"]})
 
@@ -218,11 +229,12 @@ async def mail_buttons_input(msg: Message, state: FSMContext):
 
 
 @dp.callback_query_handler(mail_action.filter(action="delete_mail"))
-async def delete_greeting(cb: CallbackQuery, callback_data: dict):
-    mail = mails_db.get(int(callback_data["id"]))
-    callback_data["id"] = mail.bot
+async def delete_mail(cb: CallbackQuery, callback_data: dict, state: FSMContext):
+    mail = await safe_get_mail(cb.from_user.id, int(callback_data["id"]), cb.id)
+    if not mail:
+        return
     mails_db.delete(mail.id)
-    await open_mail_list(cb, callback_data)
+    await open_mail_list(cb, {"id": mail.bot}, state)
 
 
 async def mail_schedule_menu(uid: int, mail_id: int, msg_id: int):
@@ -233,40 +245,39 @@ async def mail_schedule_menu(uid: int, mail_id: int, msg_id: int):
 ♻️Час видалення: {mail.del_dt.strftime(models.DT_FORMAT) if mail.del_dt else 'немає'}</i>",
         reply_markup=kb.gen_schedule_menu(mail)
     )
-    try:
-        await bot.delete_message(
-            uid,
-            msg_id
-        )
-    except MessageCantBeDeleted:
-        pass
+    await safe_del_msg(uid, msg_id)
 
 
 @dp.callback_query_handler(mail_action.filter(action="schedule"), state="*")
 async def mail_schedule_menu_cb(cb: CallbackQuery, callback_data: dict, state: FSMContext):
+    mail = await safe_get_mail(cb.from_user.id, int(callback_data["id"]), cb.id)
+    if not mail:
+        return
     await state.set_state(None)
     await mail_schedule_menu(cb.from_user.id, int(callback_data["id"]), cb.message.message_id)
 
 
 @dp.callback_query_handler(mail_action.filter(action="edit_send_dt"))
 async def edit_send_dt(cb: CallbackQuery, callback_data: dict, state: FSMContext):
+    mail = await safe_get_mail(cb.from_user.id, int(callback_data["id"]), cb.id)
+    if not mail:
+        return
     msg = await cb.message.answer(
         "Введіть дату та час у форматі <i>[H:M d.m.Y]</i>\nПриклад: <i>16:20 12.05.2023</i>",
         reply_markup=gen_cancel(mail_action.new(callback_data["id"], "schedule"))
     )
     await state.set_state(states.InputStateGroup.mail_send_dt)
     await state.set_data({"mail_id": int(callback_data["id"]), "msg_id": msg.message_id})
-    try:
-        await cb.message.delete()
-    except MessageCantBeDeleted:
-        pass
+    await safe_del_msg(cb.from_user.id, cb.message.message_id)
 
 
 @dp.message_handler(content_types=ContentTypes.TEXT, state=states.InputStateGroup.mail_send_dt)
 async def edit_send_dt(msg: Message, state: FSMContext):
-    state_data = await state.get_data()
-    mail = mails_db.get(state_data["mail_id"])
     await msg.delete()
+    state_data = await state.get_data()
+    mail = await safe_get_mail(msg.from_user.id, state_data["mail_id"])
+    if not mail:
+        return
     try:
         mail.send_dt = datetime.strptime(msg.text, models.DT_FORMAT)
     except ValueError:
@@ -284,7 +295,9 @@ async def edit_send_dt(msg: Message, state: FSMContext):
 
 @dp.callback_query_handler(mail_action.filter(action="del_send_dt"))
 async def del_send_dt(cb: CallbackQuery, callback_data: dict):
-    mail = mails_db.get(int(callback_data["id"]))
+    mail = await safe_get_mail(cb.from_user.id, int(callback_data["id"]), cb.id)
+    if not mail:
+        return
     mail.send_dt = None
     mails_db.update(mail)
     await mail_schedule_menu(cb.from_user.id, mail.id, cb.message.message_id)
@@ -292,7 +305,9 @@ async def del_send_dt(cb: CallbackQuery, callback_data: dict):
 
 @dp.callback_query_handler(mail_action.filter(action="edit_del_dt"))
 async def edit_del_dt(cb: CallbackQuery, callback_data: dict, state: FSMContext):
-    mail = mails_db.get(int(callback_data["id"]))
+    mail = await safe_get_mail(cb.from_user.id, int(callback_data["id"]), cb.id)
+    if not mail:
+        return
     bot_dc = bots_db.get(mail.bot)
     if bot_dc.premium <= 0:
         await cb.answer("⭐️Лише для преміум ботів")
@@ -303,17 +318,16 @@ async def edit_del_dt(cb: CallbackQuery, callback_data: dict, state: FSMContext)
     )
     await state.set_state(states.InputStateGroup.mail_del_dt)
     await state.set_data({"mail_id": int(callback_data["id"]), "msg_id": msg.message_id})
-    try:
-        await cb.message.delete()
-    except MessageCantBeDeleted:
-        pass
+    await safe_del_msg(cb.from_user.id, cb.message.message_id)
 
 
 @dp.message_handler(content_types=ContentTypes.TEXT, state=states.InputStateGroup.mail_del_dt)
 async def edit_del_dt(msg: Message, state: FSMContext):
-    state_data = await state.get_data()
-    mail = mails_db.get(state_data["mail_id"])
     await msg.delete()
+    state_data = await state.get_data()
+    mail = await safe_get_mail(msg.from_user.id, state_data["mail_id"])
+    if not mail:
+        return
     try:
         mail.del_dt = datetime.strptime(msg.text, models.DT_FORMAT)
     except ValueError:
@@ -331,7 +345,9 @@ async def edit_del_dt(msg: Message, state: FSMContext):
 
 @dp.callback_query_handler(mail_action.filter(action="del_del_dt"))
 async def del_del_dt(cb: CallbackQuery, callback_data: dict):
-    mail = mails_db.get(int(callback_data["id"]))
+    mail = await safe_get_mail(cb.from_user.id, int(callback_data["id"]), cb.id)
+    if not mail:
+        return
     mail.del_dt = None
     mails_db.update(mail)
     await mail_schedule_menu(cb.from_user.id, mail.id, cb.message.message_id)
@@ -339,7 +355,9 @@ async def del_del_dt(cb: CallbackQuery, callback_data: dict):
 
 @dp.callback_query_handler(mail_action.filter(action="sendout"))
 async def sendout(cb: CallbackQuery, callback_data: dict):
-    mail = mails_db.get(int(callback_data["id"]))
+    mail = await safe_get_mail(cb.from_user.id, int(callback_data["id"]), cb.id)
+    if not mail:
+        return
     bot_dc = bots_db.get(mail.bot)
     mails_db.delete(mail.id)
     await cb.message.answer(
@@ -349,10 +367,7 @@ async def sendout(cb: CallbackQuery, callback_data: dict):
             "mails"
         ))
     )
-    try:
-        await cb.message.delete()
-    except MessageCantBeDeleted:
-        pass
+    await safe_del_msg(cb.from_user.id, cb.message.message_id)
     sent_num, blocked_num, error_num = await gig.send_mail(manager.bot_dict[bot_dc.token][0], mail)
     await cb.message.answer(
         f"Розсилка {hex(mail.id*1234)} закінчена\nНадіслано: {sent_num}\nЗаблоковано:{blocked_num}\nПомилка:{error_num}",
